BackEnd/requirements.txt
amqp==5.2.0
anyio==4.3.0
asgiref==3.7.2
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
Django==5.0.3
django-compat==1.0.15
django-cors-headers==4.3.1
django-environ==0.11.2
django-jalali==6.2.1
django-jalali-date==1.1.2
django-oauth-toolkit==2.3.0
django-phonenumber-field==7.3.0
django-rest-swagger==2.2.0
django-tgbot==2.0.0
django4-background-tasks==1.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
environs==11.0.0
exceptiongroup==1.2.0
fuzzywuzzy==0.18.0
google-api-core==2.18.0
google-api-python-client==2.125.0
google-apps-meet==0.1.6
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
jalali_core==1.0.0
jdatetime==5.0.0
Jinja2==3.1.3
jwcrypto==1.5.6
kombu==5.3.7
Levenshtein==0.25.1
MarkupSafe==2.1.5
marshmallow==3.21.1
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.0
phonenumbers==8.13.32
pillow==10.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
pyTelegramBotAPI==4.17.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
python-telegram-bot==21.1
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
setuptools==69.5.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
telebot==0.0.5
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
